requirements.txt
astroid==2.4.1
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
email-validator==1.1.1
flake8==3.8.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.2
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
